requirements.txt
aiogram==3.2.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
yadisk==1.2.15
alembic==1.12.1